ENH: Default backend with __ua_covert__ with coerce=True
