+++ scipy/ndimage/tests/test_fourier.py
@@ -7,13 +7,6 @@ import pytest
 
 from scipy import ndimage
 
-from .util_backend_test import _BackendWithConverter
-from scipy.ndimage._backend import set_global_backend
-
-
-# Run the test suite with ndimage uarray backend
-set_global_backend(_BackendWithConverter, coerce=True, try_last=False)
-
 
 class TestNdimageFourier:
