+++ scipy/ndimage/tests/test_measurements.py
@@ -11,13 +11,6 @@ import scipy.ndimage as ndimage
 
 from . import types
 
-from .util_backend_test import _BackendWithConverter
-from scipy.ndimage._backend import set_global_backend
-
-
-# Run the test suite with ndimage uarray backend
-set_global_backend(_BackendWithConverter, coerce=True, try_last=False)
-
 
 class Test_measurements_stats:
     """ndimage._measurements._stats() is a utility used by other functions."""
