+++ scipy/ndimage/tests/test_morphology.py
@@ -8,13 +8,6 @@ from scipy import ndimage
 
 from . import types
 
-from .util_backend_test import _BackendWithConverter
-from scipy.ndimage._backend import set_global_backend
-
-
-# Run the test suite with ndimage uarray backend
-set_global_backend(_BackendWithConverter, coerce=True, try_last=False)
-
 
 class TestNdimageMorphology:
